Remove reshape when writing reals to arrays
When using separate_realisations to get realisations for single
candidates and the remaining set of conditionals, I used reshape when
writing realisations into the array holding realisations for estimation.
This is unnecessary if one uses the notation [i_1:i_2, ] not
[i_1:i_2, 0] when writing an array into another array. The latter will
give a broadcasting error if ones tries to write an array of shape (N,)
into it instead of something with a shape (N, 1), i.e., a non-singleton
array.

Since, _separate_realisations returns non-singleton arrays, i.e., with
shape (N, 1), we would need to do a reshape if we used the [i_1:i_2, 0]
notation. So I changed the notation to [i_1:i_2, ] everywhere such that
we don't require the reshape. Python takes care of everything as long
as both arrays have the same dimensionality along axis 1.