Fix bug: don't use minstats table if it's too small
Sequential max stats re-uses the table from the last min stats round too
save time on creating a new table. This doesn't work if the user
requested a higher number of permutations for the sequential max stats
than for the min stats.

The code should thus check if we have a sufficient number of
permutations in the min stats table and if not, a new table needs to be
created for the sequential max stats.