Add save/load functionality (stub)
Add functionality to write results dictionaries/arrays to disk
(for example, as returned by network_analysis classes like
Multivariate_te). Python does not support saving of dictionaries,
this has to be done using JSON, which in turn can not handle
numpy arrays. The load/save functions should be able to handle
this. Both functions are able to handle dictionaries, numpy arrays,
and scalar values.

Currently the functions are not handling arbitrary dictionaries. This
needs to be added in the future. There is some problem with data types
within lists, somehow Multivariate_te mixes native Python integers
and numpy integers and this causes troubles when using JSON. I need to
look into this more.