Change estimator setting in the Estimator class
Replace the old classmethods that set the estimator for the 'estimate'
method. Classmethods change the class and thus all instances of this
class. This was not the desired behavior because this way we can't have
two estimator instances with different estimator methods at the same
time.

The new implementation changes just the instance method at runtime, not
the class. It uses the 'types' module to bind the method to the class.

Add unit testing for the new implementation.